requirements.txt
streamlit==1.5.0
numpy==1.19.5
pandas==1.3.3
tensorflow==2.6.0
opencv-python==4.5.3.56
matplotlib==3.4.3
seaborn==0.11.2
Pillow==8.3.2